Refactor: PostCard 컴포넌트 리팩토링 #2
