+++ components/pages/main/Feed/PostsList/PostModal.tsx
@@ -2,7 +2,7 @@
 import { Dispatch, SetStateAction } from 'react';
 import Image from 'next/image';
 import { Comment, Post } from '@/types/types';
-import FeedPostModal from '@/components/ui/PostCard/FeedPostModal';
+import FeedPostModal from '@/components/pages/main/Feed/PostsList/FeedPostModal';
 import FlexBox from '../../../../ui/FlexBox';
 import Modal from '../../../../ui/Modal';
